moneycounter/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.1
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.8
langchain-core==1.2.8
langchain-ollama==1.0.1
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
MarkupSafe==3.0.3
more-itertools==10.8.0
networkx==3.6.1
numpy==2.3.5
ollama==0.6.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
tenacity==9.1.2
thefuzz==0.22.1
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
vine==5.1.0
wcwidth==0.5.3
xxhash==3.6.0
zstandard==0.25.0
dj-database-url
psycopg2-binary
redis